[FEAT] UseCase 추가 CreateEmptyUiDishItemUseCase : 리스트 초기화 시 필요 CreateUiDishItemUseCase : 로드한 데이터로 UiDishItem 도메인 클래스를 생성하는 유스케이스

+++ app/src/main/java/com/woowahan/android10/deliverbanchan/utils/TextBindingAdapter.kt
@@ -26,7 +26,7 @@ fun applySaleTextView(view: TextView, nPrice: Int, sPrice: Int) {
     if (nPrice == 0) {
         view.isVisible = false
     } else {
-        val sale = ((nPrice - sPrice) / nPrice * 100).toInt()
+        val sale = ((nPrice - sPrice) / nPrice.toDouble() * 100).toInt()
         val saleStr = view.context.getString(R.string.format_sale_rate, sale)
         view.isVisible = true
         view.text = saleStr
